move image loading for home page
