+++ app/page.tsx
@@ -1,21 +1,9 @@
 import Link from 'next/link'
-import { pullImageInfoFromDb } from '../actions'
-import { ToggleTheme } from './ui/ToggleTheme'
+// import { pullImageInfoFromDb } from '../actions'
+import ToggleTheme from './ui/ToggleTheme'
+import ImageView from './imageviewer/ImageView'
 
 export default async function Home (): Promise<JSX.Element> {
-  const imageInfoArray = await pullImageInfoFromDb()
-  let imagePost: JSX.Element[] = [<></>]
-  if (imageInfoArray.length !== 0) {
-    imagePost = imageInfoArray.map((fileObj, idx) => {
-      return (
-        <div key={idx}>
-          <h1>{fileObj.imagename}</h1>
-          <img src={`../assets/gallery/${fileObj.imagename}`}></img>
-          <p>{fileObj.description}</p>
-        </div>
-      )
-    })
-  }
   return (
     <main>
       <p>
@@ -24,7 +12,7 @@ export default async function Home (): Promise<JSX.Element> {
         <Link href='/dashboard'>dashboard</Link> -
         <Link href='login'>login</Link>
       </p>
-      {imageInfoArray.length > 0 ? imagePost : <p>No images to display.</p>}
+      <ImageView />
     </main>
   )
 }
